Move date range component to separate module
