+++ src/app.js
@@ -1,16 +1,13 @@
 import React, { Component } from 'react';
-import DatePicker from 'react-datepicker';
-import "react-datepicker/dist/react-datepicker.css";
 import Select from 'react-select';
 import ArticleList from './components/article-list';
+import DateRange from './components/date-range';
 import UserForm from './components/user-form';
 
 
 class App extends Component {
     state = {
-        selected: null,
-        startDate: null,
-        endDate: null
+        selected: null
     }
     render() {
         return (
@@ -24,40 +21,9 @@ class App extends Component {
                 <ArticleList
                     articles={this.props.articles}
                 />
-                <DatePicker
-                    selected={this.state.startDate}
-                    selectsStart
-                    startDate={this.state.startDate}
-                    endDate={this.state.endDate}
-                    onChange={this.handleChangeStart}
-                />
-                <DatePicker
-                    selected={this.state.endDate}
-                    selectsEnd
-                    startDate={this.state.startDate}
-                    endDate={this.state.endDate}
-                    onChange={this.handleChangeEnd}
-                />
-                <span>{this.dateRange}</span>
+                <DateRange />
             </div>
         );
-    }
-    handleSelectChange = (selected) => this.setState({ selected })
-    handleChangeStart = (date) => this.setState({ startDate: date })
-    handleChangeEnd = (date) => this.setState({ endDate: date })
-    get dateRange() {
-        if (!this.state.startDate || !this.state.endDate) {
-            return null
-        }
-        const dateDiffMillis = this.state.endDate - this.state.startDate
-        const millisInSecond = 1000
-        const secondsInMinute = 60
-        const minutesInHour = 60
-        const hoursInDay = 24
-        const dateDiffDays = dateDiffMillis / (
-            millisInSecond * secondsInMinute * minutesInHour * hoursInDay
-        )
-        return `${dateDiffDays} ${dateDiffDays === 1 ? 'day' : 'days'}`
     }
     get options() {
         return this.props.articles.map(article => ({
